refactor: start 페이지와 mypage header가 들어간 layout 생성 및 적용
